Moved cloneRepo to its own file fix  #15
